fixed a bug in the nocase sfv/filename thingy - should really be case insensitive now. fixed a bug with slowest/fastest user, if the speed was *very* high (bug id 221) default config changed slighly, but no settings was changed.

+++ zipscript/include/objects.h
@@ -108,8 +108,8 @@ struct misc {
 	char		progress_bar[15];
 	short int	release_type;
 	unsigned char	write_log;
-	short int	fastest_user[2];
-	short int	slowest_user[2];
+	long		fastest_user[2];
+	long		slowest_user[2];
 };
 
 struct VARS {
